+++ tests/testthat/test_chainof5parts.R
@@ -227,7 +227,7 @@ test_that("chainof5parts", {
   expect_true(file.exists(rn2[1]))
   TSFILE = read.csv(rn2[1])
   expect_that(nrow(TSFILE),equals(1150))
-  expect_equal(ncol(TSFILE), 11)
+  expect_equal(ncol(TSFILE), 12)
   expect_equal(length(unique(TSFILE$class_id)), 10)
   #GGIR
   suppressWarnings(GGIR(mode = c(2,3,4,5), datadir = fn, outputdir = getwd(),